requirements/common.txt
dj-database-url==0.4.2
Django==2.0.2
psycopg2==2.7.4
djangorestframework==3.7.7
pygal==2.4.0
pytz==2018.3